added u2 updated

+++ modules/apim_common/manifests/init.pp
@@ -148,12 +148,14 @@ class apim_common inherits apim_common::params {
   }
 
   # delete the wso2-updates-folder
-  exec { "delete-u2-update-pack":
-    command     => "rm -rf /root/.wso2-updates",
-    path        => "/bin/",
-    onlyif      => "/usr/bin/test -d /root/.wso2-updates",
-    subscribe   => Exec["stop-server"],
-    refreshonly => true,
+  if $enable_u2_updaes {
+    exec { "delete-u2-update-pack":
+      command     => "rm -rf /root/.wso2-updates",
+      path        => "/bin/",
+      onlyif      => "/usr/bin/test -d /root/.wso2-updates",
+      subscribe   => Exec["stop-server"],
+      refreshonly => true,
+    }
   }
 
   # Delete cert file
